Move to market metering point namespace
