+++ source/Energinet.DataHub.MeteringPoints.Domain/MeteringPoints/MarketMeteringPoints/EnergySupplierDetailsChanged.cs
@@ -12,23 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 using NodaTime;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints
+namespace Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.MarketMeteringPoints
 {
-    public class EnergySupplierDetails : ValueObject
+    public class EnergySupplierDetailsChanged : DomainEventBase
     {
-        public EnergySupplierDetails(Instant startOfSupply)
+        public EnergySupplierDetailsChanged(Guid meteringPointId, Instant startOfSupply)
         {
+            MeteringPointId = meteringPointId;
             StartOfSupply = startOfSupply;
         }
 
-        public Instant StartOfSupply { get; }
+        public Guid MeteringPointId { get; }
 
-        public static EnergySupplierDetails Create(Instant startOfSupply)
-        {
-            return new EnergySupplierDetails(startOfSupply);
-        }
+        public Instant StartOfSupply { get; }
     }
 }
